feat: 로그인 세션 연동 + 예약 승인제 전환·취소 모달·마감 캘린더·만나서결제 추가
- 로그인 세션(LoginMemberDto)에서 회원 식별하도록 연동, memberId 타입 Integer로 통일
- 결제완료→관리자 승인→예약확정→이용완료로 상태 흐름 변경(CONFIRMED 추가)
- 예약취소 사유 입력을 alert→모달(select+기타 직접입력)로 변경
- 예약 날짜를 캘린더로 교체, 숙박은 이미 예약된 날 마감(빨강) 처리(서버 이중 체크)
- 관광지·맛집은 예약금만 결제하는 만나서결제 흐름 추가(place_type 컬럼)
- 예약 관련 화면 반응형 대응

+++ src/main/java/com/gnagnoohc/travel/reservation/service/KakaoPayService.java
@@ -53,7 +53,7 @@ public class KakaoPayService {
         return res;
     }
 
-    public KakaoApproveResponse approve(String tid, String orderId, Long memberId, String pgToken) {
+    public KakaoApproveResponse approve(String tid, String orderId, Integer memberId, String pgToken) {
         Map<String, Object> body = new HashMap<>();
         body.put("cid", cid);
         body.put("tid", tid);
